creaded abstracted command layer, implemented slack and telegram bot instances. WIP - missing job commands
